requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# Web UI
streamlit>=1.28.0

# HuggingFace utilities
huggingface-hub>=0.16.0
